Rename class files to match their class declarations in JS_Mode
