+++ JS8_UI/mainwindow.h
@@ -101,12 +101,12 @@
 #include "JS8_Main/qt_helpers.h"
 #include "JS8_Main/revision_utils.h"
 #include "JS8_Main/varicode.h"
+#include "JS8_Mode/DecodedText.h"
 #include "JS8_Mode/Decoder.h"
 #include "JS8_Mode/Detector.h"
 #include "JS8_Mode/JS8.h"
 #include "JS8_Mode/JS8Submode.h"
 #include "JS8_Mode/Modulator.h"
-#include "JS8_Mode/decodedtext.h"
 #include "JS8_Network/NetworkAccessManager.h"
 #include "JS8_Network/PSKReporter.h"
 #include "JS8_Network/SpotClient.h"
